Extract Configlet class
