+++ contribution/generator/src/TrackData/PracticeExercise.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\TrackData;
 
+use App\Configlet;
 use App\TrackData\CanonicalData;
 use App\TrackData\CanonicalData\TestCase;
 use App\TrackData\Exercise;
@@ -11,7 +12,6 @@ use RuntimeException;
 
 class PracticeExercise implements Exercise
 {
-    private string $pathToConfiglet = '';
     private string $pathToPracticeExercises = '';
     private string $pathToExercise = '';
     private string $pathToCanonicalData = '';
@@ -20,7 +20,6 @@ class PracticeExercise implements Exercise
         private string $trackRoot,
         private string $exerciseSlug,
     ) {
-        $this->pathToConfiglet = $trackRoot . '/bin/configlet';
         $this->pathToPracticeExercises = $trackRoot . '/exercises/practice/';
         $this->pathToExercise =
             $this->pathToPracticeExercises . $this->exerciseSlug;
@@ -33,7 +32,6 @@ class PracticeExercise implements Exercise
 
     public function canonicalData(): CanonicalData
     {
-        $this->ensureConfigletCanBeUsed();
         $this->ensurePracticeExerciseCanBeUsed();
         $this->pathToCachedCanonicalDataFromConfiglet();
         $this->ensurePathToCanonicalDataCanBeUsed();
@@ -69,27 +67,12 @@ class PracticeExercise implements Exercise
         ), $rawData);
     }
 
-    private function ensureConfigletCanBeUsed(): void
-    {
-        if (
-            !(
-                is_executable($this->pathToConfiglet)
-                && is_file($this->pathToConfiglet)
-            )
-        ) {
-            throw new RuntimeException(
-                'configlet not found. Run the generator from track root.'
-                . ' Fetch configlet and create exercise with configlet first!'
-            );
-        }
-    }
-
     private function ensurePracticeExerciseCanBeUsed(): void
     {
         if (
             !(
-                is_writable($this->pathToExercise)
-                && is_dir($this->pathToExercise)
+                \is_writable($this->pathToExercise)
+                && \is_dir($this->pathToExercise)
             )
         ) {
             throw new RuntimeException(
@@ -101,39 +84,10 @@ class PracticeExercise implements Exercise
 
     private function pathToCachedCanonicalDataFromConfiglet(): void
     {
-        /*
-        When running configlet with detailed output (-v d) and a command that
-        requires problem specification data (e.g. info), it prints the location
-        of the cache as the first line. To avoid an HTTP call, use the offline
-        mode (-o).
-
-        Pipe the output through 'head' to get the first line only, then 'cut'
-        the 5th field to get the path only.
-
-        configlet may fail when there is no cached data (offline mode), which
-        tells us, that the exercise hasn't been generated before (the cache is
-        required for that, too). So BASH must use `-eo pipefail` to get the
-        failure code back.
-        */
-        $command = 'bash -c \'set -eo pipefail; '
-            . $this->pathToConfiglet
-            . ' -v d -t '
-            . $this->trackRoot
-            . ' info -o | head -1 | cut -d " " -f 5\''
-            ;
-        $resultCode = 1;
-
-        $configletCache = \exec(command: $command, result_code: $resultCode);
-        if ($configletCache === false || $resultCode !== 0) {
-            throw new RuntimeException(
-                '"configlet" could not provide cached canonical data.'
-                . ' Create exercise with configlet first!'
-            );
-        }
-
+        $configlet = new Configlet($this->trackRoot);
         $this->pathToCanonicalData = \sprintf(
             '%1$s/exercises/%2$s/canonical-data.json',
-            $configletCache,
+            $configlet->cachePath(),
             $this->exerciseSlug
         );
     }
@@ -142,8 +96,8 @@ class PracticeExercise implements Exercise
     {
         if (
             !(
-                is_readable($this->pathToCanonicalData)
-                && is_file($this->pathToCanonicalData)
+                \is_readable($this->pathToCanonicalData)
+                && \is_file($this->pathToCanonicalData)
             )
         ) {
             throw new RuntimeException(
